+++ main.tf
@@ -16,21 +16,27 @@
 
 locals {
   jenkins_metadata = {
-    bitnami-base-password  = "${local.jenkins_password}"
+    bitnami-base-password  = local.jenkins_password
     status-uptime-deadline = 420
-    startup-script         = "${data.template_file.jenkins_startup_script.rendered}"
+    startup-script         = data.template_file.jenkins_startup_script.rendered
   }
 
-  jenkins_password                = "${coalesce(var.jenkins_initial_password, random_string.jenkins_password.result)}"
-  jenkins_startup_script_template = "${file("${path.module}/templates/jenkins_startup_script.sh.tpl")}"
-  jenkins_username                = "user"
-
+  jenkins_password = coalesce(
+    var.jenkins_initial_password,
+    random_string.jenkins_password.result,
+  )
+  jenkins_startup_script_template = file(
+    "${path.module}/templates/jenkins_startup_script.sh.tpl"
+  )
+  jenkins_username            = "user"
   jenkins_workers_project_url = "https://www.googleapis.com/compute/v1/projects/${var.jenkins_workers_project_id}"
 
-  jenkins_workers_startup_script = <<EOF
+  jenkins_workers_startup_script  = <<EOF
 ${data.template_file.jenkins_workers_agent_startup_script.rendered}
 ${var.jenkins_workers_startup_script}
 EOF
+  jenkins_instance_access_configs = var.jenkins_instance_private ? [] : var.jenkins_instance_access_configs
+  jenkins_workers_access_configs  = var.jenkins_workers_private ? [] : var.jenkins_workers_access_configs
 }
 
 resource "random_string" "jenkins_password" {
@@ -39,82 +45,93 @@ resource "random_string" "jenkins_password" {
 }
 
 data "google_compute_image" "jenkins" {
-  name    = "${var.jenkins_boot_disk_source_image}"
-  project = "${var.jenkins_boot_disk_source_image_project}"
+  name    = var.jenkins_boot_disk_source_image
+  project = var.jenkins_boot_disk_source_image_project
 }
 
 data "google_compute_image" "jenkins_worker" {
-  name    = "${var.jenkins_workers_boot_disk_source_image}"
-  project = "${var.jenkins_workers_boot_disk_source_image_project}"
+  name    = var.jenkins_workers_boot_disk_source_image
+  project = var.jenkins_workers_boot_disk_source_image_project
 }
 
 data "template_file" "jenkins_workers_agent_startup_script" {
-  template = "${file("${path.module}/templates/jenkins_workers_agent_startup_script.sh.tpl")}"
+  template = file(
+    "${path.module}/templates/jenkins_workers_agent_startup_script.sh.tpl",
+  )
 
-  vars {
-    jenkins_username = "${local.jenkins_username}"
-    jenkins_password = "${local.jenkins_password}"
+  vars = {
+    jenkins_username = local.jenkins_username
+    jenkins_password = local.jenkins_password
   }
 }
 
 data "template_file" "jenkins_startup_script" {
-  template = "${local.jenkins_startup_script_template}"
+  template = local.jenkins_startup_script_template
 
-  vars {
-    jenkins_username                               = "${local.jenkins_username}"
-    jenkins_password                               = "${local.jenkins_password}"
-    jenkins_workers_project_id                     = "${var.jenkins_workers_project_id}"
-    jenkins_workers_instance_cap                   = "${var.jenkins_workers_instance_cap}"
-    jenkins_workers_description                    = "${var.jenkins_workers_description}"
-    jenkins_workers_name_prefix                    = "${var.jenkins_workers_name_prefix}"
+  vars = {
+    jenkins_username                               = local.jenkins_username
+    jenkins_password                               = local.jenkins_password
+    jenkins_workers_project_id                     = var.jenkins_workers_project_id
+    jenkins_workers_instance_cap                   = var.jenkins_workers_instance_cap
+    jenkins_workers_description                    = var.jenkins_workers_description
+    jenkins_workers_name_prefix                    = var.jenkins_workers_name_prefix
     jenkins_workers_region                         = "${local.jenkins_workers_project_url}/regions/${var.jenkins_workers_region}"
     jenkins_workers_zone                           = "${local.jenkins_workers_project_url}/zones/${var.jenkins_workers_zone}"
     jenkins_workers_machine_type                   = "${local.jenkins_workers_project_url}/zones/${var.jenkins_workers_zone}/machineTypes/${var.jenkins_workers_machine_type}"
     jenkins_workers_startup_script                 = "${local.jenkins_workers_startup_script}"
-    jenkins_workers_preemptible                    = "${var.jenkins_workers_preemptible ? "true" : "false"}"
-    jenkins_workers_min_cpu_platform               = "${var.jenkins_workers_min_cpu_platform}"
-    jenkins_workers_labels                         = "${join(",", var.jenkins_workers_labels)}"
-    jenkins_workers_run_as_user                    = "${var.jenkins_workers_run_as_user}"
+    jenkins_workers_preemptible                    = var.jenkins_workers_preemptible ? "true" : "false"
+    jenkins_workers_min_cpu_platform               = var.jenkins_workers_min_cpu_platform
+    jenkins_workers_labels                         = join(",", var.jenkins_workers_labels)
+    jenkins_workers_run_as_user                    = var.jenkins_workers_run_as_user
     jenkins_workers_boot_disk_type                 = "${local.jenkins_workers_project_url}/zones/${var.jenkins_workers_zone}/diskTypes/${var.jenkins_workers_boot_disk_type}"
-    jenkins_workers_boot_disk_source_image         = "${data.google_compute_image.jenkins_worker.self_link}"
-    jenkins_workers_boot_disk_source_image_project = "${var.jenkins_workers_boot_disk_source_image_project}"
-    jenkins_workers_network                        = "${var.jenkins_workers_network}"
-    jenkins_workers_subnetwork                     = "${var.jenkins_workers_subnetwork}"
-    jenkins_workers_network_tags                   = "${join(",", var.jenkins_workers_network_tags)}"
-    jenkins_workers_service_account_email          = "${var.jenkins_workers_service_account_email}"
-    jenkins_workers_retention_time_minutes         = "${var.jenkins_workers_retention_time_minutes}"
-    jenkins_workers_launch_timeout_seconds         = "${var.jenkins_workers_launch_timeout_seconds}"
-    jenkins_workers_boot_disk_size_gb              = "${var.jenkins_workers_boot_disk_size_gb}"
-    jenkins_workers_num_executors                  = "${var.jenkins_workers_num_executors}"
-    jobs_as_b64_json                               = "${base64encode(jsonencode(var.jenkins_jobs))}"
+    jenkins_workers_boot_disk_source_image         = data.google_compute_image.jenkins_worker.self_link
+    jenkins_workers_boot_disk_source_image_project = var.jenkins_workers_boot_disk_source_image_project
+    jenkins_workers_network                        = var.jenkins_workers_network
+    jenkins_workers_subnetwork                     = var.jenkins_workers_subnetwork
+    jenkins_workers_network_tags                   = join(",", var.jenkins_workers_network_tags)
+    jenkins_workers_service_account_email          = var.jenkins_workers_service_account_email
+    jenkins_workers_retention_time_minutes         = var.jenkins_workers_retention_time_minutes
+    jenkins_workers_launch_timeout_seconds         = var.jenkins_workers_launch_timeout_seconds
+    jenkins_workers_boot_disk_size_gb              = var.jenkins_workers_boot_disk_size_gb
+    jenkins_workers_num_executors                  = var.jenkins_workers_num_executors
+    jenkins_workers_use_internal_address           = var.jenkins_workers_private ? "true" : "false"
+    jenkins_workers_external_address               = var.jenkins_workers_private ? "false" : "true"
+    jobs_as_b64_json                               = base64encode(jsonencode(var.jenkins_jobs))
   }
 }
 
 resource "google_compute_instance" "jenkins" {
-  project      = "${var.project_id}"
-  name         = "${var.jenkins_instance_name}"
-  machine_type = "${var.jenkins_instance_machine_type}"
-  zone         = "${var.jenkins_instance_zone}"
+  project      = var.project_id
+  name         = var.jenkins_instance_name
+  machine_type = var.jenkins_instance_machine_type
+  zone         = var.jenkins_instance_zone
 
-  tags = "${var.jenkins_instance_tags}"
+  tags = var.jenkins_instance_tags
 
   boot_disk {
     initialize_params {
-      image = "${data.google_compute_image.jenkins.self_link}"
+      image = data.google_compute_image.jenkins.self_link
     }
   }
 
   network_interface {
-    subnetwork         = "${var.jenkins_instance_subnetwork}"
-    subnetwork_project = "${var.project_id}"
+    subnetwork         = var.jenkins_instance_subnetwork
+    subnetwork_project = var.project_id
 
-    access_config {}
+    dynamic "access_config" {
+      for_each = local.jenkins_instance_access_configs
+      content {
+        nat_ip                 = lookup(access_config, "nat_ip", null)
+        public_ptr_domain_name = lookup(access_config, "public_ptr_domain_name", null)
+        network_tier           = lookup(access_config, "network_tier", null)
+      }
+    }
   }
 
-  metadata = "${merge(local.jenkins_metadata, var.jenkins_instance_additional_metadata)}"
+  metadata = merge(local.jenkins_metadata, var.jenkins_instance_additional_metadata)
 
   service_account {
-    email = "${google_service_account.jenkins.email}"
+    email = google_service_account.jenkins.email
 
     scopes = [
       "https://www.googleapis.com/auth/cloud-platform",
@@ -127,5 +144,5 @@ resource "null_resource" "wait_for_jenkins_configuration" {
     command = "${path.module}/scripts/wait-for-jenkins.sh ${var.project_id} ${var.jenkins_instance_zone} ${var.jenkins_instance_name}"
   }
 
-  depends_on = ["google_compute_instance.jenkins"]
+  depends_on = [google_compute_instance.jenkins]
 }