+++ test/fixtures/simple_example/network.tf
@@ -26,8 +26,7 @@ resource "random_string" "suffix" {
 }
 
 provider "google" {
-  credentials = "${file(local.credentials_path)}"
-  project     = "${var.project_id}"
+  project = var.project_id
 }
 
 resource "google_compute_network" "main" {
@@ -38,14 +37,14 @@ resource "google_compute_network" "main" {
 resource "google_compute_subnetwork" "subnetwork" {
   name          = "cft-jenkins-test-${local.example_name}-${random_string.suffix.result}"
   ip_cidr_range = "10.0.0.0/21"
-  region        = "${var.region}"
-  network       = "${google_compute_network.main.self_link}"
+  region        = var.region
+  network       = google_compute_network.main.self_link
 }
 
 resource "google_compute_firewall" "ssh" {
   name    = "cft-jenkins-test-${local.example_name}-ssh-access"
-  network = "${google_compute_network.main.self_link}"
-  project = "${var.project_id}"
+  network = google_compute_network.main.self_link
+  project = var.project_id
 
   allow {
     protocol = "tcp"